update modules/ai: mod        4 files

+++ modules/ai/src/main/java/com/bytedesk/ai/robot/RobotLlm.java
@@ -2,7 +2,7 @@
  * @Author: jackning 270580156@qq.com
  * @Date: 2024-06-05 10:02:51
  * @LastEditors: jackning 270580156@qq.com
- * @LastEditTime: 2025-04-19 14:33:08
+ * @LastEditTime: 2025-04-22 10:39:23
  * @Description: bytedesk.com https://github.com/Bytedesk/bytedesk
  *   Please be aware of the BSL license restrictions before installing Bytedesk IM – 
  *  selling, reselling, or hosting Bytedesk IM as a service is a breach of the terms and automatically terminates your rights under the license.
@@ -74,11 +74,11 @@ public class RobotLlm {
     // 温度越低，回答结果越固定，随机性越低
     @Builder.Default
     @Column(name = "llm_temperature")
-    private float temperature = 0.0f;
+    private Double temperature = 0.0;  // 从float改为Double
 
     @Builder.Default
     @Column(name = "llm_top_p")
-    private float topP = 0.7f;
+    private Double topP = 0.7;  // 从float改为Double
 
     @Builder.Default
     @Column(name = "llm_prompt", columnDefinition = TypeConsts.COLUMN_TYPE_TEXT)
